refactor(screen): draw the compile card from its own class
The mark that pulses in a corner while a pack compiles, and the words
beside it, were drawn by PackChain, the one file of the frame importing
a font and the HUD's extractor, and reading six things of the chain to
do it. CompileCard draws them off what the chain now answers: the words
of the moment, how long ago the last family finished compiling, and
which load the chain came from, which keys the card's own fade-in clock
in place of a field on the chain. What the corner shows is unchanged.
